Fixed all warnings

+++ src/CLITerminalWindow.cpp
@@ -65,8 +65,8 @@ void CLITerminalWindow::render() {
                     init_color(15-i, palette[i].r * (1000/255), palette[i].g * (1000/255), palette[i].b * (1000/255));
             lastPaletteChecksum = checksum;
         }
-        for (int y = 0; y < screen.size(); y++) {
-            for (int x = 0; x < screen[y].size(); x++) {
+        for (int y = 0; (unsigned)y < screen.size(); y++) {
+            for (int x = 0; (unsigned)x < screen[y].size(); x++) {
                 move(y, x);
                 addch(screen[y][x] | COLOR_PAIR(colors[y][x]));
             }
